+++ src/Settings/PreferencesWindow.cpp
@@ -30,6 +30,7 @@
 #include <QStackedWidget>
 #include <QTreeWidget>
 #include <QVBoxLayout>
+#include <generated/SettingsHelper.hpp>
 
 PreferencesWindow::PreferencesWindow(QWidget *parent) : QMainWindow(parent)
 {
@@ -190,7 +191,7 @@ void PreferencesWindow::switchToPage(QWidget *page, bool force)
     // ask for saving changes or not if not force
     if (!force)
     {
-        auto current = dynamic_cast<PreferencesPage *>(stackedWidget->currentWidget());
+        auto current = qobject_cast<PreferencesPage *>(stackedWidget->currentWidget());
         if (current != nullptr && !current->aboutToExit())
             return;
     }
@@ -201,7 +202,7 @@ void PreferencesWindow::switchToPage(QWidget *page, bool force)
     // switch if everything is OK
     stackedWidget->setCurrentWidget(page);
 
-    auto preferencesPage = dynamic_cast<PreferencesPage *>(page);
+    auto preferencesPage = qobject_cast<PreferencesPage *>(page);
     if (preferencesPage != nullptr)
     {
         menuTree->setCurrentItem(pageTreeItem[preferencesPage]);
@@ -278,8 +279,8 @@ PreferencesPage *PreferencesWindow::getPageWidget(const QString &pagePath)
 
 void PreferencesWindow::closeEvent(QCloseEvent *event)
 {
-    auto current = dynamic_cast<PreferencesPage *>(stackedWidget->currentWidget());
-    if (current != nullptr && !current->aboutToExit())
+    auto current = qobject_cast<PreferencesPage *>(stackedWidget->currentWidget());
+    if (!SettingsHelper::isForceClose() && current != nullptr && !current->aboutToExit())
         event->ignore();
     else
         event->accept();