+++ src/Settings/SettingsManager.cpp
@@ -115,6 +115,8 @@ void SettingsManager::init()
 
 void SettingsManager::deinit()
 {
+
+	LOG_INFO("Started saving settings");
     QSettings setting(configFileLocation[0], QSettings::IniFormat);
     setting.clear(); // Otherwise SettingsManager::remove won't work
 
@@ -140,6 +142,8 @@ void SettingsManager::deinit()
     delete cur;
     delete def;
     cur = def = nullptr;
+
+	LOG_INFO("All Settings have been saved");
 }
 
 QVariant SettingsManager::get(QString key, bool alwaysDefault)