+++ src/appwindow.cpp
@@ -91,8 +91,23 @@ AppWindow::AppWindow(bool noHotExit, QWidget *parent) : QMainWindow(parent), ui(
     applySettings();
     onSettingsApplied("");
 
-    if (!noHotExit && SettingsHelper::isHotExitEnable())
+    do
     {
+        if (noHotExit || (!SettingsHelper::isForceClose() && !SettingsHelper::isHotExitEnable()))
+            break;
+
+        SettingsHelper::setForceClose(false);
+
+        if (!SettingsHelper::isHotExitEnable())
+        {
+            auto res = QMessageBox::question(
+                this, "Hot Exit",
+                "In the last session, CP Editor was abnormally killed, do you want to restore the last session?",
+                QMessageBox::Yes | QMessageBox::No);
+            if (res == QMessageBox::No)
+                break;
+        }
+
         int length = SettingsHelper::getHotExitTabCount();
 
         QProgressDialog progress(this);
@@ -131,7 +146,7 @@ AppWindow::AppWindow(bool noHotExit, QWidget *parent) : QMainWindow(parent), ui(
 
         if (currentIndex >= 0 && currentIndex < ui->tabWidget->count())
             ui->tabWidget->setCurrentIndex(currentIndex);
-    }
+    } while (false);
 }
 
 AppWindow::AppWindow(int depth, bool cpp, bool java, bool python, bool noHotExit, const QStringList &paths,
@@ -172,6 +187,7 @@ AppWindow::AppWindow(bool cpp, bool java, bool python, bool noHotExit, int numbe
 
 AppWindow::~AppWindow()
 {
+    LOG_INFO("Destruction started");
     saveSettings();
     Extensions::EditorTheme::release();
     delete ui;
@@ -188,6 +204,7 @@ AppWindow::~AppWindow()
     delete findReplaceDialog;
 
     SettingsManager::deinit();
+    LOG_INFO("Destruction finished");
 }
 
 /******************* PUBLIC METHODS ***********************/
@@ -384,7 +401,7 @@ void AppWindow::openTab(const QString &path)
         auto fileInfo = QFileInfo(path);
         for (int t = 0; t < ui->tabWidget->count(); t++)
         {
-            auto tmp = dynamic_cast<MainWindow *>(ui->tabWidget->widget(t));
+            auto tmp = qobject_cast<MainWindow *>(ui->tabWidget->widget(t));
             if (fileInfo == QFileInfo(tmp->getFilePath()))
             {
                 ui->tabWidget->setCurrentIndex(t);
@@ -537,27 +554,29 @@ void AppWindow::saveEditorStatus(bool loadFromFile)
 
 bool AppWindow::quit()
 {
-    bool ret = false;
-    if (SettingsHelper::isHotExitEnable())
+    if (preferencesWindow->isVisible() && !preferencesWindow->close())
+        return false;
+    if (SettingsHelper::isHotExitEnable() || SettingsHelper::isForceClose())
     {
         LOG_INFO("quit() with hotexit");
         SettingsHelper::setHotExitLoadFromFile(false);
         saveEditorStatus(false);
-        ret = true;
     }
     else
     {
         LOG_INFO("quit() called without hotExit");
         on_actionClose_All_triggered();
-        ret = ui->tabWidget->count() == 0;
+        if (ui->tabWidget->count() >= 1)
+        {
+            LOG_INFO("Closing is cancelled");
+            return false;
+        }
     }
-    if (ret && preferencesWindow->isVisible())
-        ret = preferencesWindow->close();
     // The tray icon is considered as a visible window, if it is not hidden, even if the app window is closed,
     // the application won't exit.
-    if (ret)
-        trayIcon->hide();
-    return ret;
+    trayIcon->hide();
+    LOG_INFO("All preparations for closing are finished");
+    return true;
 }
 
 int AppWindow::getNewUntitledIndex()
@@ -782,6 +801,12 @@ void AppWindow::onReceivedMessage(quint32 instanceId, QByteArray message)
 }
 
 #undef FROMJSON
+
+bool AppWindow::forceClose()
+{
+    SettingsHelper::setForceClose(true);
+    return close();
+}
 
 void AppWindow::onTabCloseRequested(int index)
 {
@@ -1461,7 +1486,7 @@ MainWindow *AppWindow::currentWindow()
     {
         return nullptr;
     }
-    return dynamic_cast<MainWindow *>(ui->tabWidget->widget(current));
+    return qobject_cast<MainWindow *>(ui->tabWidget->widget(current));
 }
 
 void AppWindow::reAttachLanguageServer(MainWindow *window)
@@ -1504,7 +1529,7 @@ MainWindow *AppWindow::windowAt(int index)
     {
         return nullptr;
     }
-    return dynamic_cast<MainWindow *>(ui->tabWidget->widget(index));
+    return qobject_cast<MainWindow *>(ui->tabWidget->widget(index));
 }
 
 void AppWindow::on_actionShow_Logs_triggered()